Update to Lucene 9.3.0 (#4043)
(cherry picked from commit bab2e59bb15a4e6e0a5c926f5ea2d7cc4318365b)

+++ server/src/main/java/org/opensearch/index/fielddata/ordinals/MultiOrdinals.java
@@ -231,8 +231,8 @@ public class MultiOrdinals extends Ordinals {
         }
 
         @Override
-        public long docValueCount() {
-            return currentEndOffset - currentOffset;
+        public int docValueCount() {
+            return Math.toIntExact(currentEndOffset - currentOffset);
         }
     }
 }
